Factor out geo-restriction logic
No functional change.

Impetus is to make the logic re-usable for comment-blocking.
