+++ ui/redux/selectors/claims.js
@@ -15,6 +15,7 @@ import {
 } from 'util/claim';
 import * as CLAIM from 'constants/claim';
 import { INTERNAL_TAGS } from 'constants/tags';
+import { getGeoRestrictionForClaim } from 'util/geoRestriction';
 
 type State = { claims: any, user: UserState };
 
@@ -858,34 +859,6 @@ export const selectGeoRestrictionForUri = createCachedSelector(
   selectGeoBlockLists,
   selectUserLocale,
   (claim, geoBlockLists, locale: LocaleInfo) => {
-    if (locale && geoBlockLists && claim) {
-      const claimId: ?string = claim.claim_id;
-      const channelId: ?string = getChannelIdFromClaim(claim);
-
-      let geoConfig: ?GeoConfig;
-
-      // --- livestreams
-      if (isStreamPlaceholderClaim(claim) && geoBlockLists.livestreams) {
-        geoConfig = geoBlockLists.livestreams[channelId] || geoBlockLists.livestreams[claimId];
-      }
-      // --- videos (a.k.a everything else)
-      else if (geoBlockLists.videos) {
-        geoConfig = geoBlockLists.videos[channelId] || geoBlockLists.videos[claimId];
-      }
-
-      if (geoConfig) {
-        const specials = geoConfig.specials || [];
-        const countries = geoConfig.countries || [];
-        const continents = geoConfig.continents || [];
-
-        return (
-          specials.find((x: GeoRestriction) => x.id === 'EU-ONLY' && locale.is_eu_member) ||
-          countries.find((x: GeoRestriction) => x.id === locale.country) ||
-          continents.find((x: GeoRestriction) => x.id === locale.continent)
-        );
-      }
-    }
-
-    return null;
+    return getGeoRestrictionForClaim(claim, locale, geoBlockLists);
   }
 )((state, uri) => String(uri));
